feat: MySQL에 데이터 삽입/삭제 시 Elasticsearch에도 같이 반영되게 만들기
변경 사항

- CreateProductRequestDto에서 가격 타입을 Long에서 Int로 변경
- Product 및 ProductDocument에서 가격 타입을 Long에서 Int로 변경
- createProduct 메서드에서 @RequestBody 어노테이션 추가
- deleteProduct 메서드에서 @PathVariable 어노테이션 추가
- ProductService에서 Elasticsearch에 제품 저장 로직 추가

영향 범위

- 제품 생성 및 삭제 API
- 데이터베이스 및 Elasticsearch 연동

관련 이슈 : main

+++ src/main/kotlin/com/dino/dino_coupang_api_server/product/domain/Product.kt
@@ -19,7 +19,7 @@ class Product(
     @Column(columnDefinition = "TEXT")
     var description: String, // 에디터가 HTML로 상품 설명 저장
 
-    var price: Long,
+    var price: Int,
 
     var rating: Double,
